requirements.txt
django
urllib3
djangorestframework
markdown
django-filter
psycopg2-binary